+++ templates/wizard.php
@@ -479,7 +479,7 @@
                 </div>
                 <div style="display: table-row;">
                     <div style="display: table-cell; width: 50%;">
-                        <fieldset class="flexibleDatasetMode">
+                        <fieldset class="flexibleDatasetMode" hidden>
                             <legend><?php p($l->t('Dataset structure')); ?></legend>
                             <label>
                                 <input type="radio" name="wizardDatasetMode" value="legacy" checked>